requirements.txt
flask
Flask-SQLAlchemy
flask-login
python-dateutil
flask-mail